Moved rep status widget code to platform/user folder
