+++ src/platform/user/widgets/representative-status/components/States/Auth.jsx
