+++ src/platform/user/widgets/representative-status/stylesheet.scss
