+++ src/platform/user/widgets/representative-status/tests/fixtures/mocks/loggedInUser.json
