+++ src/platform/user/widgets/representative-status/tests/formatContactInfo.unit.spec.jsx
