+++ src/platform/user/widgets/representative-status/tests/representativeTypeMap.unit.spec.js
