+++ src/platform/user/widgets/representative-status/utilities/formatContactInfo.js
