+++ src/platform/user/widgets/representative-status/utilities/representativeTypeMap.js
